+++ devolo_home_control_api/publisher/updater.py
@@ -249,11 +249,13 @@ class Updater:
     def _device_events(self, message: dict):
         """ If an operation was not successful, we need to correct our internal state. """
         properties = {"properties": message.get("properties").get("property.value.new")}
-        properties.get("properties")["uid"] = properties.get("properties").get("widgetElementUID")
-        if get_device_type_from_element_uid(properties.get("properties").get("widgetElementUID")) == "devolo.BinarySwitch":
-            properties.get("properties")["property.name"] = "state"
-            properties.get("properties")["property.value.new"] = int(properties.get("properties").get("data"))
-        self.update(properties)
+        if type(properties.get("properties")) is dict:
+            properties['properties']['uid'] = properties.get("properties").get("widgetElementUID")
+            if get_device_type_from_element_uid(properties['properties']['uid']) == "devolo.BinarySwitch":
+                # TODO: Check, if we need to handle other device types than BinarySwitch on unsuccessful operations
+                properties['properties']['property.name'] = "state"
+                properties['properties']['property.value.new'] = int(properties.get("properties").get("data"))
+            self.update(properties)
 
     def _device_online_state(self, message: dict):
         """ Update the device's online state. """