+++ tests/mocks/mock_siren.py
@@ -6,6 +6,7 @@ import requests
 from devolo_home_control_api.devices.zwave import Zwave
 from devolo_home_control_api.properties.binary_sensor_property import BinarySensorProperty
 from devolo_home_control_api.properties.multi_level_switch_property import MultiLevelSwitchProperty
+from devolo_home_control_api.properties.settings_property import SettingsProperty
 
 from .mock_gateway import MockGateway
 
@@ -38,5 +39,12 @@ def siren(device_uid: str) -> Zwave:
                                  session=session,
                                  element_uid=f'devolo.SirenMultiLevelSwitch:{device_uid}',
                                  state=test_data.get("devices").get("siren").get("state"))
+
+    device.settings_property = {}
+    device.settings_property['muted'] = \
+        SettingsProperty(gateway=gateway,
+                         session=session,
+                         element_uid=f'bas.{device_uid}',
+                         value=test_data.get("devices").get("siren").get("muted"))
 
     return device