+++ docs/CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.12.0] - 2020/06/09
+
+### Added
+
+- Support for devolo Siren
+
+### Changed
+
+- **BREAKING**: The naming of some setting properties changed
+
+### Fixed
+
+- Searching for devices now also works, if multiple devices have the same name.
+- Websocket messages in unexpected format are handled gracefully.
+
 ## [v0.11.0] - 2020/05/26
 
 ### Added